Create Dataset
We built our dataset from scratch, we collected 1546 images
related to religious topics and categorized it to three classes.
We droped all files smaller than 10 k, resized the images
to 180*180mm. We discovered several duplicate rows in the
dataset but kept them because of the lack of images in this
topic. we removed any image has extension not related to the
extensions (jpeg,jpg, bmp, png], then distributed the data as
follows: validation: 0.2, training: 0.8.
Our main dataset topic is religion. We distributed the files
to 3 sub-folders presenting 3 classes (Islamic, Christian, and
Jewish). The type of labels is inferred, the type of label mode
is categorical, and the color mode is RGB.